Move evaluation to evaluate
